wip: keep going if already removed

+++ lib/Ravada/VM/KVM.pm
@@ -361,6 +361,7 @@ sub search_volume_re($self,$pattern,$refresh=0) {
 sub remove_file($self,@files) {
     for my $file (@files) {
         if ($self->is_local) {
+            next if ! -e $file;
             unlink $file or die "$! $file";
             next;
         }
